MT#9617 Apply debian-changes patch as sipwise-debian-changes.patch
As identified via the according binary build of
https://jenkins.mgm.sipwise.com/job/heartbeat-sources/ build
number 206 all the Makefiles get rebuilt if we don't apply the
debian-changes that has been used so far (see e.g. working builds
like #208 and #211).

Without this patch we get:

| [...]
| missing --run aclocal-1.7
| [....]
| missing --run automake-1.7
| [...]
| missing --run autoconf
| [...]
| /bin/bash ./config.status --recheck
| [...]
| configure: creating ./config.status
| [...]
| /bin/bash ./config.status Makefile
| config.status: creating Makefile

during the build process (tools like aclocal-1.7 and automake-1.7
aren't even available at all). The resulting heartbeat binary
doesn't use /usr/lib64 in its library path then, therefore
failing with:

| /usr/lib64/heartbeat/heartbeat: error while loading shared libraries: libstonith.so.1: cannot open shared object file: No such file or directory

One possible workaround for this would be to run:

| echo /usr/lib64 > /etc/ld.so.conf.d/heartbeat.conf; ldconfig

but to avoid possible further hidden problems with this let's
apply sipwise-debian-changes.patch. This patch is the
result of actually using the source from

| dget -u http://archive.debian.org/debian/pool/main/h/heartbeat/heartbeat_2.1.3-6lenny4.dsc

as its base since the very beginning of this repository.
dpkg-source automatically recorded those changes for us as
debian-changes file (see commit id f94588851c9f549146de26). We
want to be able to have a clean repository state for a decent
workflow with git and quilt.  Therefore instead of using this
patch implicitly, make this patch an explicit one.