Update patch set 1
Patch Set 1:

> You knows it better, no ideas here.
 > LGTM, +1 to show to other guys.
 > P.S. I see no difference in lintian output actually (30 errors):
 > 
 > review: https://jenkins.mgm.sipwise.com/job/heartbeat-binaries/54/testReport/
 > 
 > trunk: https://jenkins.mgm.sipwise.com/job/heartbeat-binaries/50/testReport/
 > 
 > So, what we are fixing here?

Sorry, didn't want to add more to the commit message, because given that this is a licensing issue, it might expose Sipwise to some liability, and I'm not sure how transparent we are with this kind of issues.

The problem here is that heartbeat is GPL-2+ and it is currently linking against OpenSSL transitively via cURL, so that's a license violation (to some at least), given that using the GnuTLS variant should be functionally equivalent here it should be the safest option. Also cURL is not used any longer in the latest upstream, here it is only used by the drac3 plugin, which I'm not even sure we actually use at all.

Patch-set: 1