Update patch set 1
Patch Set 1:

As I said up 2 you here, binaries were built successfully so we can try it in production. BTW is the heartbeat debian package comming from Sipwise or we just fork it (rebuilding to apply some fixes)? If we forked it , is it legal to change such dependencies?

Patch-set: 1