Update patch set 1
Patch Set 1:

> As I said up 2 you here, binaries were built successfully so we can
 > try it in production.

I tried locally and everything looked fine, but given that cURL is only used by a hardware specific plugin, it would be hard for me to actually test the actual change. :/

 > BTW is the heartbeat debian package comming
 > from Sipwise or we just fork it (rebuilding to apply some fixes)?

It is a complete fork, we have a heartbeat-2 package that has not been synced with Debian for a very long time, version is 2.1.3 forked from a Debian lenny stable from 2009. Current version in unstable is 3.0.5. :/

 > If we forked it , is it legal to change such dependencies?

Depends if by legal you mean valid (which it is, as the dependencies should be feature equivalent), or law-related (which I don't see any problem with either, because not doing the change is what can cause legal issues). Hope that clarifies. :)

Patch-set: 1
Label: Code-Review=0