TT#26264 Cleanup heartbeat service support
Move init script into /usr/share/heartbeat so that it is not easily
accessible as a proper sysvinit script. It did not use lsb/init-functions
and as such was not redirecting to the systemd service, making the script
dangerous to have laying around as it could mess up with our systemd
support.

Change-Id: Ia317b719cf23933523cb3f945d701015384e39cc