TT#36043 Switch Requires=monit.service to Wants=monit.service for heartbeat
Heartbeat doesn't depends on monit anymore since mr6.2+.
HA uses ngcp-service and ngcp-service calls systemctl directly to
wait for services to be up and ready. It allows to handle services
dependencies properly. ngcp-service also calls 'monit monitor/unmonitor'
while it should work well with Wants=monit.service, in the worst case
the monit on active node will show 'Not monitored' for already running
services. It is a price for keeping HA running on monit service restart.

The price is acceptable here as we are going to remove monit anyway and
'ngcp-service summary' output is already reliable here for all the
corner cases.

Change-Id: I9060ca63ac0e7aceadf045f6f4465811f3c595ad