TT#36043 Remove monit.service dependency information from service files
We do not require monit in any way now. Listing it as a dependency and
declaring an After ordering means that on reboot, heartbeat is always
shut down first, it will then try to shut down all NGCP services, and
meanwhile monit will notice and try to restart them, making heartbeat
not be able to shut down cleanly.

Change-Id: I454c7714159b92c9237f2530ba8ba82316481478