TT#101063 Switch from python 2 to 3
Switch python shebang to python3, and switch python scripts to python3
by using the 2to3 conversion tool. Update Build-Depends, and remove now
redundant non-dev python dependency as python3-dev takes that role in
addition to pulling in libpython3-dev.

This is needed for the upgrade to Debian bullseye, which has removed
most of python 2.

Change-Id: If86477b23591369fe2fb494ccf430f60425d754f