TT#101063 Remove autogenerated files
This is safe as dpkg-source will ignore deleted files during build. Then
we run autoreconf during the build, which makes sure that we regenerate
any file during the build process, and get the latest versions, which
should handle support for Python 3 among others, needed for the upgrade
to Debian bullseye.

Change-Id: Ic55c8dbc0912e90a263617bdbb9c97142c138b9c